Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/textarea-selection-while-parsing-expected.txt



FAIL Value setter while parsing textarea children should move selection{Start,End} to the end assert_equals: expected 3 but got 0
